Models (#63)
* added wrappers for the models and some tests

* fixing prepare data for singletons

* updated slightly the doc

* added docstring in test_model.py

* update docstring

* added TrainModel class

* some more tests and small fixes

* added additionnal tests for cameras

* small fixes

* minor fix

* removing print

* added  automated data generation in the sim

* updated API

* fixing test for replay camera

* added wrappers for the models and some tests

* fixing prepare data for singletons

* updated slightly the doc

* added docstring in test_model.py

* update docstring

* added TrainModel class

* some more tests and small fixes

* added additionnal tests for cameras

* small fixes

* minor fix

* removing print

* added  automated data generation in the sim

* updated API

* fixing test for replay camera

* updating api

* fixing tests on github

* updating api

* fixing tests on github

* fixing datagenerator and integration with training !

* trained our first model !

* small fix

* added data augm (will need to refactor / dig further)

* quick fix

+++ autopylot/cameras/dummy.py
@@ -32,4 +32,4 @@ class Dummy:
 
     def update(self):
         """Create a dummy image."""
-        self.memory["image"] = np.random.random(self.shape) * 255
+        self.memory["image"] = (np.random.random(self.shape) * 255).astype(np.int32)
